backend/requirements.txt
Flask==1.1.1
gunicorn==20.0.4
natsort==7.0.1
psycopg2==2.8.6
pylint
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml
whitenoise==5.2.0
wrapt==1.11.2
torch @ https://download.pytorch.org/whl/cpu/torch-1.5.1%2Bcpu-cp38-cp38-linux_x86_64.whl
torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.6.1%2Bcpu-cp38-cp38-linux_x86_64.whl
pillow
gunicorn==20.0.4
Flask-Cors